Updates for Run History Detail

+++ LucentDb.Domain/Model/ValidationResponse.cs
@@ -10,7 +10,7 @@ namespace LucentDb.Domain.Model
         public bool IsValid { get; set; }
         public string ResultMessage { get; set; }
         public string RunLog { get; set; }
-        public double Duration { get; set; }
+        public decimal Duration { get; set; }
         public DateTime RunDateTime { get; set; }
     }
 }
